More refactoring: move various query binders to their own classes
Also move various classes to their own files
