+++ modules/backend/src/main/scala/utils/binders/NamespaceExtractor.scala
