+++ modules/core/app/services/search/SearchParams.scala
@@ -2,7 +2,9 @@ package services.search
 
 import defines.EntityType
 import play.api.libs.json.{Format, Json, Writes}
+import play.api.mvc.QueryStringBindable
 import utils.EnumUtils
+import utils.binders._
 
 
 case object SearchField extends Enumeration {
@@ -81,6 +83,15 @@ object BoundingBox {
   }
 
   implicit val writes: Writes[BoundingBox] = Json.writes[BoundingBox]
+
+  implicit def bboxQueryBinder(implicit stringBinder: QueryStringBindable[String]): QueryStringBindable[BoundingBox] =
+    new QueryStringBindable[BoundingBox] {
+
+      override def bind(key: String, params: Map[String, Seq[String]]): Option[Either[String, BoundingBox]] =
+        params.get(key).flatMap(_.headOption).map(BoundingBox.fromString)
+
+      override def unbind(key: String, value: BoundingBox): String = stringBinder.unbind(key, value.toString)
+    }
 }
 
 case class LatLng(
@@ -108,6 +119,15 @@ object LatLng {
   }
 
   implicit val writes: Writes[LatLng] = Json.writes[LatLng]
+
+  implicit def latLngQueryBinder(implicit stringBinder: QueryStringBindable[String]): QueryStringBindable[LatLng] =
+    new QueryStringBindable[LatLng] {
+
+      override def bind(key: String, params: Map[String, Seq[String]]): Option[Either[String, LatLng]] =
+        params.get(key).flatMap(_.headOption).map(LatLng.fromString)
+
+      override def unbind(key: String, value: LatLng): String = stringBinder.unbind(key, value.toString)
+    }
 }
 
 /**
@@ -144,4 +164,68 @@ object SearchParams {
   def empty: SearchParams = SearchParams()
 
   implicit val writes: Writes[SearchParams] = Json.writes[SearchParams]
+
+  implicit def _queryBinder(
+    implicit seqStrBinder: QueryStringBindable[Seq[String]]): QueryStringBindable[SearchParams] with NamespaceExtractor = new QueryStringBindable[SearchParams] with NamespaceExtractor {
+
+    private implicit val sortBinder: QueryStringBindable[SearchSort.Value] = queryStringBinder(SearchSort)
+
+    // Backwards compatibility for old Solr-based `sort parameters. We fall
+    // back to these if none of the new sort values are valid.
+    def bindOldSort(key: String, params: Map[String, Seq[String]]): Option[SearchSort.Value] = {
+      val oldSortMap: Map[String, SearchSort.Value] = Map(
+        "identifier.asc" -> SearchSort.Id,
+        "isParent.desc,identifier.asc" -> SearchSort.Id,
+        "score.desc" -> SearchSort.Score,
+        "name_sort.asc" -> SearchSort.Name,
+        "lastUpdated.desc" -> SearchSort.DateNewest,
+        "countryCode.asc" -> SearchSort.Country,
+        "repositoryName.asc" -> SearchSort.Holder,
+        "geodist().asc" -> SearchSort.Location,
+        "charCount.desc" -> SearchSort.Detail,
+        "childCount.desc" -> SearchSort.ChildCount
+      )
+      params
+        .get(key)
+        .map(_.flatMap(oldSortMap.get).headOption)
+        .getOrElse(Option.empty)
+    }
+
+
+    override def bind(key: String, params: Map[String, Seq[String]]): Option[Either[String, SearchParams]] = {
+      val namespace: String = ns(key)
+      val rawParams = SearchParams(
+        bindOr(namespace + QUERY, params, Option.empty[String]).filter(_.trim.nonEmpty),
+        bindOr(namespace + SORT, params, bindOldSort(namespace + SORT, params)),
+        bindOr(namespace + ENTITY, params, Seq.empty[EntityType.Value])(tolerantSeqBinder(queryStringBinder(EntityType))),
+        bindOr(namespace + FIELD, params, Seq.empty[SearchField.Value])(tolerantSeqBinder(queryStringBinder(SearchField))),
+        bindOr(namespace + FACET, params, Seq.empty[String]),
+        bindOr(namespace + EXCLUDE, params, Seq.empty[String]),
+        bindOr(namespace + FILTERS, params, Seq.empty[String]),
+        bindOr(namespace + BBOX, params, Option.empty[BoundingBox]),
+        bindOr(namespace + LATLNG, params, Option.empty[LatLng])
+      )
+      // NB: Sorting by location is invalid without a valid `latlng` parameter
+      val checked = if (rawParams.sort.contains(SearchSort.Location) && rawParams.latLng.isEmpty)
+        rawParams.copy(sort = None)
+      else rawParams
+
+      Some(Right(checked))
+    }
+
+    override def unbind(key: String, params: SearchParams): String =
+      utils.http.joinQueryString(toParams(params, ns(key)).distinct)
+
+    private def toParams(p: SearchParams, ns: String = ""): Seq[(String, String)] = {
+      p.query.map(q => ns + QUERY -> q).toSeq ++
+        p.sort.map(s => ns + SORT -> s.toString).toSeq ++
+        p.entities.map(e => ns + ENTITY -> e.toString) ++
+        p.fields.map(f => ns + FIELD -> f.toString) ++
+        p.facets.map(f => ns + FACET -> f) ++
+        p.excludes.map(e => ns + EXCLUDE -> e) ++
+        p.filters.map(f => ns + FILTERS -> f) ++
+        p.bbox.map { box => ns + BBOX -> box.toString }.toSeq ++
+        p.latLng.map { latLng => ns + LATLNG -> latLng.toString }.toSeq
+    }
+  }
 }
